+++ ffupdater/src/main/java/de/marmaro/krt/ffupdater/app/App.kt
@@ -12,7 +12,8 @@ enum class App(appDetail: BaseApp) {
     BRAVE(Brave()),
     ICERAVEN(Iceraven()),
     BROMITE(Bromite()),
-    VIVALDI(Vivaldi());
+    VIVALDI(Vivaldi()),
+    STYX(Styx());
 
     val detail: BaseApp = appDetail
 }